barlabels {plotrix}	R Documentation

Label the bars on a barplot

Description

Displays labels on a plot, usually a bar plot.

Usage

 barlabels(xpos,ypos,labels=NULL,cex=1,prop=0.5,miny=0,offset=0,...)

Arguments

xpos
A vector, matrix or data frame of x positions for the labels.

ypos
A vector, matrix or data frame of y positions for the labels.

labels
The labels to display. Defaults to the values of y.

cex
Relative size of the labels. See text.

prop
The proportion of ypos at which to place the labels. Defaults to 0.5 (the middle).

miny
The minimum value at which to display labels.

offset
Amount to horizontally offset successive labels in case of vertical overlaps.

...
Extra arguments passed to boxed.labels.

Details

barlabels places labels on a plot at horizontal positions xpos and vertical positions ypos * pos. The typical use of this function is to place labels on bars, by default in the middle of the bars.

To put labels just over the tops of the bars, set prop to 1 and add a constant amount to ypos.

Value

nil

Author(s)

Jim Lemon

See Also

boxed.labels

Examples

 heights<-c(14,20,9,31,17)
 barpos<-barplot(heights,main="A redundant bar plot")
 # show the usual value labels on the bars
 barlabels(barpos,heights)
 # now with stacked bars and offsets
 heights<-matrix(sample(c(1,2,10,15),20,TRUE),ncol=4)
 barpos<-barplot(heights,main="Another redundant bar plot")
 barlabels(barpos,heights,offset=0.1)
 # finally use barp for the plot
 barpos<-barp(heights,main="A third and final bar plot",col=2:6,
  names.arg=paste("Day",1:4))
 barlabels(barpos$x,barpos$y,matrix(LETTERS[1:5],nrow=5,ncol=4))

[Package plotrix version 3.5-2 Index]