binciWu {plotrix}	R Documentation

Upper binomial confidence limit

Description

Calculates an upper binomial confidence limit using the Wilson approximation.

Usage

 binciWu(x,n,alpha=0.05,trueval=NA,na.rm=TRUE)

Arguments

x
The number of successes or failures for which the CI is to be calculated.

n
The number of trials as above.

alpha
The desired coverage - 0.05 produces 95 percent coverage

trueval
The value representing the outcome of interest for the CI.

na.rm
Argument needed to make this work

Details

binciWu calculates the upper binomial confidence limit for the given number of successes and trials. It is mainly to allow binomial confidence limits to be calculated in the brkdnNest function, which is why the upper and lower CIs are calculated separately.

Value

The upper binomial confidence interval

Author(s)

Jim Lemon

See Also

binciWl

Examples

 binciWu(5,42)

[Package plotrix version 3.5-2 Index]