box.heresy {plotrix}	R Documentation

Display a sort of box plot

Description

box.heresy displays a box plot in which a symbol represents a measure of central tendency, a surrounding box that represents an "inner" measure of dispersion (e.g. standard error) and whiskers that represent an "outer" measure of dispersion (e.g. standard deviation). The function is pretty basic at this time and will probably change a bit.

Usage

 box.heresy(x,y,uinner,linner=uinner,ulim,llim=ulim,intervals=FALSE,
 arrow.cap=NA,pch=22,main="",xlab="",ylab="",xaxlab=NA,col="white",do.first=NULL,...)

Arguments

x,y
Vectors of numeric values representing measures of central tendency.

uinner,linner
Vectors of numeric values representing "inner" measures of dispersion.

ulim,llim
Vectors of numeric values representing "outer" measures of dispersion.

intervals
Whether the values for dispersion are intervals (TRUE) or absolute limits (FALSE).

arrow.cap
The width of the cap on the "whiskers" relative to the width of the plot. Defaults to the same width as the outer box.

pch
The symbol to be used to represent the measure(s) of central tencency in the box.

main
The title for the plot (i.e. main).

xlab,ylab
The x and y axis labels.

xaxlab
Optional labels for the boxes.

col
The fill colors for the "inner" rectangles.

do.first
An expression that will be evaluated before anything is displayed.

...
additional arguments passed to the dispersion function.

Value

nil

Author(s)

Jim Lemon - thanks to Gianni Lavaredo for the suggestion

See Also

plot, boxplot

Examples

 y<-runif(5)
 ulim<-runif(5)
 llim<-runif(5)
 uinner<-ulim/2
 linner<-llim/2
 box.heresy(y,uinner=uinner,linner=linner,ulim=ulim,llim=llim,
  intervals=TRUE,main="The heretical boxplot",
  xlab="Number of observations",ylab="Value")

[Package plotrix version 3.5-2 Index]