convert.meq {plotrix}	R Documentation

Convert milligrams/liter to milliequivalents/liter.

Description

Convert milligrams/liter of ions used in the Piper diagram to milliequivalents/liter.

Usage

 convert.meq(mgpl,parameter)

Arguments

mgpl 	 A data frame containing the concentrations of ions used in the Piper diagram in milligrams/liter.

parameter 	 The names of the ions if necessary.

Details

convert.meq does the conversion with a lookup table of molecular weights and valences.

Value

The concentration of the ions in milliequivalents.

Author(s)

Mike Cheetham

See Also

piper.diagram

[Package plotrix version 2.7 Index]