drawSectorAnnulus {plotrix}	R Documentation

Display a radial pie sector

Description

Displays a radial pie sector with optional annuli.

Usage

 drawSectorAnnulus(angle1,angle2,radius1,radius2,col,angleinc=0.03)

Arguments

angle1,angle2
Start and end angle for the sector.

radius1,radius2
Start and end of the radial extent of the annulus.

col
Color of the sector.

angleinc
The angular increment to use in drawing the arcs.

Details

drawSectorAnnulus displays a single radial pie sector. It is probably only useful when called from radial.pie.

Value

nil

Author(s)

Jim Lemon

See Also

radial.pie

[Package plotrix version 3.5-2 Index]