fill.corner {plotrix}	R Documentation

Fill a "corner" of a matrix with values

Description

Fills one corner of a matrix with the supplied values, leaving the rest filled with NAs.

Usage

 fill.corner(x,nrow,ncol)

Arguments

x
A vector of values.

nrow,ncol
The number of rows and columns in the matrix to be returned.

Details

fill.corner creates an nrow by ncol matrix and fills the lower left corner with the values supplied in x. If there are more values in x than cells in the matrix, only the first nrow*ncol values will be inserted.

Value

An nrow by ncol matrix containing the values in x.

Author(s)

Jim Lemon

[Package plotrix version 3.5-2 Index]