get.triprop {plotrix}	R Documentation

Enter three proportion data - usually soil textures

Description

get.triprop allows the user to enter triplets of proportions or percentages of three components such as sand, silt and clay in soils.

Usage

 get.triprop(use.percentages=FALSE,cnames=c("1st","2nd","3rd"))

Arguments

use.percentages
Logical - whether to treat the entries as percentages and scale to proportions.

cnames
column names for the resulting three column matrix.

Details

The three proportions of each row must sum to 100 or 1 within 1% or the function will warn the operator.

Value

A matrix of the components of one or more observations.

Author(s)

Jim Lemon

See Also

triax.plot, soil.texture

Examples

 if(dev.interactive()) {
  # get some proportions
  newsp<-get.triprop()
  # show the triangle
  triax.frame(main="Test triax.plot")
  # now plot the observations
  triax.points(newsp)
 }

[Package plotrix version 3.5-2 Index]