l2010 {plotrix}	R Documentation

World lightning strike data from 2010

Description

A list of two 50x100 matrices containing most of the world lightning strike data from 2010. It was produced by makeDensityMatrix from 171 file (about 3 Gb) of data consisting in two geographic coordinates for the approximate location of each recorded strike and an estimated intensity of the strike in kVA.

Usage


 data(l2010)

[Package plotrix version 3.8-2 Index]