listDepth {plotrix}	R Documentation

Find the maximum depth of a list.

Description

Descend a list and find the maximum number of levels.

Usage

 listDepth(x) 

Arguments

x 	 A list.

Details

A possibly nested list of lists is descended to determine the maximum number of levels. Currently used to set up the dimensions of a dendrite plot.

Value

The maximum number of levels in the list.

Author(s)

Jim Lemon

See Also

plot.dendrite

[Package plotrix version 2.7 Index]