p.ions {plotrix}	R Documentation

Calculate ion/total(ion) ratios.

Description

Calculates the ion/total(ions) ratios of ions used in the Piper diagram.

Usage

 p.ions(meq,short.output=TRUE)

Arguments

meq 	 A data frame containing the concentrations of ions used in the Piper diagram in milliequivalents/liter.

short.output 	 Whether to return the proportions of four ions (TRUE) or include carbonate-bicarbonate and sodium-potassium (FALSE).

Details

p.ions calculates the proportions of anions or cations to total anions or cations for the ions used in the Piper diagram.

Value

The proportion of the ions.

Author(s)

Mike Cheetham

See Also

piper.diagram

[Package plotrix version 2.7 Index]