piper.points {plotrix}	R Documentation

Display the points on a Piper diagram.

Description

Displays the points for the anions and cations in the Piper diagram.

Usage

 piper.points(ca,mg,so4,cl,ions=data.frame(ca=ca,mg=mg,so4=so4,cl=cl),sites,
  ppm=TRUE,chull=FALSE,pch=3,main="",cex.pch=1,col=NA,pch.lwd=0.7,plot.sep=0.15)

Arguments

ca,mg,so4,cl 	 Concentrations of the ions.

ions 	 data frame containing the above values.

sites 	 indices for ions.

ppm 	 Whether the concentrations are in milligrams/liter (parts per million) or milliequivalents.

chull 	 Whether to display points or polygons for the concentrations.

pch 	 Symbol to use in plotting points.

main 	 title for the diagram

cex.pch 	 Expansion for the points.

col 	 Colors for the points/polygons.

pch.lwd 	 Line widths.

plot.sep 	 The offsets for the ions.

Details

piper.points displays the points/polygons in the Piper diagram.

Value

A list containing the coordinates of the points and optionally information about the indices and colors.

Author(s)

Mike Cheetham

See Also

piper.diagram

[Package plotrix version 2.7 Index]