piper.set {plotrix}	R Documentation

Calculate ion offsets.

Description

Calculates the offsets for anions and cations in the Piper diagram.

Usage

 piper.set(plot.sep=0.15)

Arguments

plot.sep 	 The offsets of the ions in user units.

Details

piper.set calculates the ion offsets in the Piper diagram.

Value

The offsets.

Author(s)

Mike Cheetham

See Also

piper.diagram

[Package plotrix version 2.7 Index]