plot.dendrite {plotrix}	R Documentation

Plot a dendrogram of a dendrite object

Description

Plot a dendrogram for two or more mutually exclusive attributes.

Usage

 ## S3 method for class 'dendrite'
plot(x,xlabels=NULL,main="",mar=c(1,0,3,0),cex=1,...)

Arguments

x
A dendrite object containing the counts of objects having combinations of mutually exclusive attributes.

xlabels
The category labels that will be displayed beneath the dendrogram.

main
The title of the plot.

mar
Margins for the plot.

cex
Character expansion for the leaves of the dendrogram.

...
Additional arguments passed to plot.

Details

plot.dendrite sets up a plot for a dendrogram. The actual plotting of the dendrogram is done by furc.

Value

nil

Author(s)

Jim Lemon

See Also

furc

Examples

 sex<-sample(c("M","F"),100,TRUE)
 hair<-c(sample(c("Blond","Black","Brown","Red"),90,TRUE),rep(NA,10))
 eye<-sample(c("Blue","Black","Brown","Green"),100,TRUE)
 charac<-data.frame(sex=sex,hair=hair,eye=eye)
 characlist<-makeDendrite(charac)
 plot.dendrite(characlist,names(charac),main="Test dendrogram",cex=0.8)

[Package plotrix version 3.5-2 Index]