plotrixInfo {plotrix}	R Documentation

Save information from a plot

Description

plotrixInfo is a list filled by some functions with an add argument.

Details

Because plotrixInfo is created in the parent environment of a function, it may be read if the same function is called with add=TRUE. This allows some plot parameters to be passed from the initial plotting function to subsequent calls that add more values.

[Package plotrix version 3.5-2 Index]