polar.plot {plotrix}	R Documentation

Plot values on a circular grid of 0 to 360 degrees.

Description

polar.plot displays a plot of radial lines, symbols or a polygon centered at the midpoint of the plot frame on a 0:360 circle. Positions are interpreted as beginning at the right and moving counterclockwise unless start specifies another starting point or clockwise is TRUE.

Usage

 polar.plot(lengths,polar.pos=NULL,labels,label.pos=NULL,
  start=0,clockwise=FALSE,rp.type="r",...)

Arguments

lengths 	 numeric data vector. Magnitudes will be represented as the radial positions of symbols, line ends or polygon vertices.

polar.pos 	 numeric vector of positions on a 0:360 degree circle. These will be converted to radians when passed to radial.plot.

labels 	 text labels to place on the periphery of the circle. This defaults to labels every 20 degrees. For no labels, pass an empty string.

label.pos 	 positions of the peripheral labels in degrees

start 	 The position for zero degrees on the plot in degrees.

clockwise 	 Whether to increase angles clockwise rather than the default counterclockwise.

rp.type 	 Whether to plot radial lines, symbols or a polygon.

... 	 additional arguments passed to radial.plot and then to plot.

Value

nil

Author(s)

Jim Lemon

See Also

radial.plot

Examples

 testlen<-c(rnorm(36)*2+5)
 testpos<-seq(0,350,by=10)
 polar.plot(testlen,testpos,main="Test Polar Plot",lwd=3,line.col=4)
 polar.plot(testlen,testpos,main="Test Clockwise Polar Plot",
  start=90,clockwise=TRUE,lwd=3,line.col=4)
 # reset the margins
 par(mar=c(5,4,4,2))

[Package plotrix version 2.7 Index]