propbrk {plotrix}	R Documentation

Calculate the proportion of specified values in a vector

Description

Calculates the proportion of values in a vector that are equal to a specified value.

Usage

 propbrk(x,trueval=TRUE,na.rm=TRUE)

Arguments

x
a character, factor or numeric vector.

trueval
the value to be matched in x.

na.rm
whether to remove NA values.

Details

propbrk calculates the proportion of values matching a specified value. It is mainly to allow proportions to be calculated in the brkdnNest function. It now discards NA values in x when calculating proportions.

Value

nil

Author(s)

Jim Lemon

See Also

brkdnNest

Examples

 propbrk(sample(LETTERS,100,TRUE),trueval="M")

[Package plotrix version 3.5-2 Index]