staxlab {plotrix}	R Documentation

Place staggered labels on an axis

Description

Places labels on an axis in a regular staggered order

Usage

 staxlab(side=1,at,labels,nlines=2,top.line=0.5,line.spacing=0.8,...)

Arguments

side 	 axis on which to place the labels, as in axis

at 	 where to place the labels in user units, as in axis

labels 	 text strings

nlines 	 How many lines to use to stagger the labels.

top.line 	 Distance from the axis to place the first line of text.

line.spacing 	 Spacing between lines of text labels.

... 	 Additional arguments to be passed to mtext.

Value

nil

Note

This function is mainly useful when either long axis labels or a large number of labels are to be placed without overlapping. It staggers the labels along the axis specified. The user may wish to increase the space beneath the plot using mar before calling staxlab. It is probably only useful on the bottom or left side of the plot.

Author(s)

Jim Lemon

See Also

mtext

Examples

 plot(rnorm(12),axes=FALSE)
 box()
 months<-c("January","February","March","April","May","June",
  "July","August","September","October","November","December")
 staxlab(1,1:12,months)

[Package plotrix version 2.7 Index]