valid.n {plotrix}	R Documentation

Find the number of valid (not NA) values

Description

Finds the number of valid (not NA) or total values in an object.

Usage

 valid.n(x,na.rm=TRUE)

Arguments

x
An object.

na.rm
Whether to count all values (FALSE) or only those not NA.

Details

valid.n finds the number of valid values of the object if na.rm=TRUE.

Value

The number of valid values or the length of the object.

Author(s)

Jim Lemon

[Package plotrix version 3.5-2 Index]