Perform a nested breakdown of numeric values

Breaks down a numeric or categorical element of a data frame by one or more categorical elements.

Usage

brkdnNest(formula,data,FUN=c("mean","sd","sd","valid.n"),label1="Overall",
  trueval=TRUE)

Arguments

formula

A formula with a numeric element of a data frame on the left and one or more categorical elements on the right.

data

A data frame containing the elements in formula.

FUN

The functions to be applied to successive breakdowns.

label1

The label to use for the overall value of the first function.

trueval

The value to use in calculating proportions or sums of a categorical response variable. See Details.

Value

A list with as many elements as there are functions in FUN. It is probably best to always specify four functions (summary measure, upper dispersion measure, lower dispersion measure and number of valid observations) even if this is redundant as in the default.

This function is similar to brkdn in the prettyR package, but is structured to be used with the barNest function. It produces one or more measures for the overall data, then the subsets of the data defined by the first variable to the right of the tilde, then the subsets defined by the first and second variable, and so on.

Details

brkdnNest performs a nested breakdown of an element of a data frame by one or more categorical elements. For each category and optional subcategories, the variable on the left of the formula is summarized as specified by the functions named in FUN.

If trueval is not NA, brkdnNest will calculate the proportion of trueval values in the response variable out of the total valid responses. If the function valid.n is the first function in FUN, the counts of the groups and subgroups will be returned.

Two specialized summary functions are defined within brkdnNest. sumbrk returns the count of values in a factor equal to trueval, and propbrk returns the proportion of values equal to trueval. Be aware that if a categorical variable is specified on the left of the formula, functions which expect numeric data such as mean should not be included in FUN.

The user should take care when specifying different summary functions. barNest expects a summary measure as the first component of the list and measures of dispersion as the second and third. If two different measures of dispersion are passed, the first must calculate the upper and the second the lower measure.

Author

Jim Lemon

See also

Examples

 brkdntest<-data.frame(Age=rnorm(100,25,10),
  Sex=factor(sample(c("M","F"),100,TRUE)),
  Marital=sample(c("M","X","S","W"),100,TRUE),
  Employ=sample(c("FT","PT","NO"),100,TRUE))
 brkdnNest(formula=Age~Sex+Marital+Employ,data=brkdntest)
#> mean 
#> Overall 24.64639 
#> 	 F 25.48129 
#> 	 	 M 24.14016 
#> 	 	 	 FT 28.81242 
#> 	 	 	 NO 18.70251 
#> 	 	 	 PT 26.33539 
#> 	 	 S 23.0874 
#> 	 	 	 FT 28.00318 
#> 	 	 	 NO 18.80666 
#> 	 	 	 PT 22.45237 
#> 	 	 W 30.73962 
#> 	 	 	 FT 21.44249 
#> 	 	 	 NO 40.4357 
#> 	 	 	 PT 33.57271 
#> 	 	 X 26.74302 
#> 	 	 	 FT 26.26994 
#> 	 	 	 NO 27.90948 
#> 	 	 	 PT 25.66082 
#> 	 M 23.906 
#> 	 	 M 24.47751 
#> 	 	 	 FT 29.30051 
#> 	 	 	 NO 27.46603 
#> 	 	 	 PT 20.17864 
#> 	 	 S 25.18825 
#> 	 	 	 FT 30.4521 
#> 	 	 	 NO 13.93655 
#> 	 	 	 PT 30.04115 
#> 	 	 W 26.9493 
#> 	 	 	 FT 28.16036 
#> 	 	 	 NO 25.81686 
#> 	 	 	 PT 26.22584 
#> 	 	 X 19.91266 
#> 	 	 	 FT 18.50095 
#> 	 	 	 NO 13.30853 
#> 	 	 	 PT 24.39116 
#> sd 
#> Overall 10.12796 
#> 	 F 8.629231 
#> 	 	 M 6.877148 
#> 	 	 	 FT 8.065114 
#> 	 	 	 NO 4.912362 
#> 	 	 	 PT 5.403243 
#> 	 	 S 8.673395 
#> 	 	 	 FT 6.018149 
#> 	 	 	 NO 9.631543 
#> 	 	 	 PT 8.98311 
#> 	 	 W 10.45648 
#> 	 	 	 FT 7.179402 
#> 	 	 	 NO 2.917489 
#> 	 	 	 PT 9.548556 
#> 	 	 X 8.456904 
#> 	 	 	 FT 8.930538 
#> 	 	 	 NO 10.57663 
#> 	 	 	 PT 8.334899 
#> 	 M 11.32376 
#> 	 	 M 9.425406 
#> 	 	 	 FT 6.03256 
#> 	 	 	 NO 15.98711 
#> 	 	 	 PT 9.03891 
#> 	 	 S 12.80534 
#> 	 	 	 FT 9.820226 
#> 	 	 	 NO 13.87821 
#> 	 	 	 PT 7.946644 
#> 	 	 W 8.484358 
#> 	 	 	 FT 6.202651 
#> 	 	 	 NO 5.97119 
#> 	 	 	 PT 19.45212 
#> 	 	 X 12.83002 
#> 	 	 	 FT 11.09681 
#> 	 	 	 NO 0.5976962 
#> 	 	 	 PT 16.64632 
#> sd 
#> Overall 10.12796 
#> 	 F 8.629231 
#> 	 	 M 6.877148 
#> 	 	 	 FT 8.065114 
#> 	 	 	 NO 4.912362 
#> 	 	 	 PT 5.403243 
#> 	 	 S 8.673395 
#> 	 	 	 FT 6.018149 
#> 	 	 	 NO 9.631543 
#> 	 	 	 PT 8.98311 
#> 	 	 W 10.45648 
#> 	 	 	 FT 7.179402 
#> 	 	 	 NO 2.917489 
#> 	 	 	 PT 9.548556 
#> 	 	 X 8.456904 
#> 	 	 	 FT 8.930538 
#> 	 	 	 NO 10.57663 
#> 	 	 	 PT 8.334899 
#> 	 M 11.32376 
#> 	 	 M 9.425406 
#> 	 	 	 FT 6.03256 
#> 	 	 	 NO 15.98711 
#> 	 	 	 PT 9.03891 
#> 	 	 S 12.80534 
#> 	 	 	 FT 9.820226 
#> 	 	 	 NO 13.87821 
#> 	 	 	 PT 7.946644 
#> 	 	 W 8.484358 
#> 	 	 	 FT 6.202651 
#> 	 	 	 NO 5.97119 
#> 	 	 	 PT 19.45212 
#> 	 	 X 12.83002 
#> 	 	 	 FT 11.09681 
#> 	 	 	 NO 0.5976962 
#> 	 	 	 PT 16.64632 
#> valid.n 
#> Overall 100 
#> 	 F 47 
#> 	 	 M 14 
#> 	 	 	 FT 3 
#> 	 	 	 NO 5 
#> 	 	 	 PT 6 
#> 	 	 S 15 
#> 	 	 	 FT 5 
#> 	 	 	 NO 5 
#> 	 	 	 PT 5 
#> 	 	 W 8 
#> 	 	 	 FT 3 
#> 	 	 	 NO 2 
#> 	 	 	 PT 3 
#> 	 	 X 10 
#> 	 	 	 FT 3 
#> 	 	 	 NO 4 
#> 	 	 	 PT 3 
#> 	 M 53 
#> 	 	 M 14 
#> 	 	 	 FT 5 
#> 	 	 	 NO 2 
#> 	 	 	 PT 7 
#> 	 	 S 16 
#> 	 	 	 FT 7 
#> 	 	 	 NO 5 
#> 	 	 	 PT 4 
#> 	 	 W 9 
#> 	 	 	 FT 4 
#> 	 	 	 NO 3 
#> 	 	 	 PT 2 
#> 	 	 X 14 
#> 	 	 	 FT 5 
#> 	 	 	 NO 3 
#> 	 	 	 PT 6 
 # show the proportion of unemployed with binomial confidence intervals
 brkdnNest(formula=Employ~Sex+Marital,data=brkdntest,
  FUN=c("propbrk","binciWu","binciWl"),trueval="NO")
#> propbrk 
#> Overall 0.29 
#> 	 F 0.3404255 
#> 	 	 M 0.3571429 
#> 	 	 S 0.3333333 
#> 	 	 W 0.25 
#> 	 	 X 0.4 
#> 	 M 0.245283 
#> 	 	 M 0.1428571 
#> 	 	 S 0.3125 
#> 	 	 W 0.3333333 
#> 	 	 X 0.2142857 
#> binciWu 
#> Overall 0.3853889 
#> 	 F 0.4832908 
#> 	 	 M 0.6123558 
#> 	 	 S 0.5828645 
#> 	 	 W 0.5907246 
#> 	 	 X 0.6873262 
#> 	 M 0.3756656 
#> 	 	 M 0.3994138 
#> 	 	 S 0.5559564 
#> 	 	 W 0.6457979 
#> 	 	 X 0.4758923 
#> binciWl 
#> Overall 0.2101484 
#> 	 F 0.2216744 
#> 	 	 M 0.1634473 
#> 	 	 S 0.1517632 
#> 	 	 W 0.07147921 
#> 	 	 X 0.1681803 
#> 	 M 0.149329 
#> 	 	 M 0.04009392 
#> 	 	 S 0.1416464 
#> 	 	 W 0.1205838 
#> 	 	 X 0.07571387