clean.args {plotrix}	R Documentation

Remove inappropriate arguments from an argument list

Description

Takes a list of arguments and eliminates those that are not appropriate for passing to a particular function (and hence would produce an error if passed).

Usage

 clean.args(argstr,fn,exclude.repeats=FALSE,exclude.other=NULL,dots.ok=TRUE)
 remove.args(argstr,fn)

Arguments

argstr 	 a named list of arguments, e.g. from ...

fn 	 a function

exclude.repeats 	 (logical) remove repeated arguments?

exclude.other 	 a character vector of names of additional arguments to remove

dots.ok 	 should "..." be allowed in the argument list?

Value

clean.args returns a list which is a copy of argstr with arguments inappropriate for fn removed; remove.args removes the arguments for fn from the list.

Author(s)

Ben Bolker

Examples

 tststr <- list(n=2,mean=0,sd=1,foo=4,bar=6) 
 clean.args(tststr,rnorm)
 try(do.call("rnorm",tststr))
 do.call("rnorm",clean.args(tststr,rnorm))
 remove.args(tststr,rnorm)
 ## add example of combining arg. lists?

[Package plotrix version 2.7 Index]