color.id {plotrix}	R Documentation

Identify closest match to a color

Description

Given a color specified as a hex string, find the closest match in the table of known (named) colors

Usage


 color.id(col)

Arguments

col
a color specified as a hex string

Details

finds the color with the minimum squared distance in RGB space

Value

the name of the closest match

Author(s)

Ben Bolker

See Also

col2rgb,colors

Examples


 color.id("#cc00cc")

[Package plotrix version 3.8-2 Index]