fan.plot {plotrix}	R Documentation

Display a fan plot

Description

Displays numerical values as the arcs of overlapping sectors.

Usage

 fan.plot(x,edges=200,radius=1,col=NULL,align.at=NULL,max.span=NULL,
  labels=NULL,labelpos=NULL,label.radius=1.2,align="left",shrink=0.02,
  main="",ticks=NULL,include.sumx=FALSE,...)

Arguments

x
Vector of numbers.

edges
The number of edges with which to draw a circle.

radius
The radius of the sectors.

col
The colors with which to fill the sectors.

align.at
Where to align the sectors (see Details).

max.span
The angle of the maximal sector in radians. The default is to scale x so that it sums to 2*pi.

labels
Labels placed around the sector arcs.

labelpos
Optional circumferential positions for the labels.

label.radius
How far away from the sectors the labels will be placed. May be a vector with a radius for each label.

align
Position of the alignment of sectors (see Details).

shrink
How much to shrink each successive sector in user units.

main
Optional title for the plot.

ticks
The number of ticks that would appear if the sectors were on a pie chart. Default is no ticks, TRUE gives the number of ticks equal to the integer sum of x, which is fairly sensible if x is a vector of integers.

include.sumx
Whether to include the sum of all x values as the largest sector.

...
Additional arguments passed to polygon.

Details

fan.plot is now a call to fanPlot.

Value

The circumferential positions of the labels in radians. These are returned in order of decreasing size of the values plotted.

Author(s)

Jim Lemon, Anupam Tyagi

[Package plotrix version 3.5-2 Index]