feather.plot {plotrix}	R Documentation

Display vectors along a horizontal reference line.

Description

Displays vectors along a line usually representing time or position.

Usage

 feather.plot(r,theta,xpos,yref=0,use.arrows=TRUE,
 col.refline="lightgray",fp.type="s",main="",xlab="",ylab="",
 xlabels=NULL,...)

Arguments

r 	 radii of vectors

theta 	 direction of vectors in radians

xpos 	 where to start each vector along the reference line

yref 	 vertical position to place the reference line

use.arrows 	 whether to put arrow heads on the ends of the vectors

col.refline 	 the color of the reference line

fp.type 	 whether to use "standard" coordinates (begin at the right and move counterclockwise) or "meteorological" coordinates (begin at the top and move clockwise) when interpreting the values of theta

main 	 the title of the plot

xlab 	 the label for the reference line

ylab 	 the label for the vertical axis

xlabels 	 optional labels for the reference line

... 	 additional arguments passed to arrows or segments

Details

This function places vectors of length r and angle theta along a reference line that may represent time or position or some other value. The user is responsible for spacing the vectors so that they do not overlap if this is desired.

Value

nil

Author(s)

Jim Lemon, Eduardo Klein

See Also

spread.labels

Examples

 feather.plot(0.6+rnorm(8)/5,seq(0,7*pi/4,by=pi/4),1:8,
  main="Standard Coordinates",xlab="Time",ylab="Value")
 if(dev.interactive()) par(ask=TRUE)
 feather.plot(0.6+rnorm(8)/5,seq(0,7*pi/4,by=pi/4),1:8,
  main="Meteorological Coordinates",xlab="Time",ylab="Value",
  fp.type="m",xlabels=TRUE)
 par(ask=FALSE)

[Package plotrix version 2.7 Index]