furc {plotrix}	R Documentation

Plot a dendrite

Description

Plot one level of a dendrogram displaying two or more mutually exclusive attributes.

Usage

 furc(x,xpos,yrange,toplevel,maxx,cex=1,col)

Arguments

x
A dendrite object containing the counts of objects having combinations of mutually exclusive attributes.

xpos
The horizontal position on the plot to display the current level of the dendrogram.

yrange
The range of values in which the current level of the dendrogram will be displayed.

toplevel
A flag for the function to know whether it is at the top level of the dendrogram or not. Do not change this argument.

maxx
The maximum xpos value to suppress staggering of the category labels. This is calculated in plot.dendrite.

cex
The character expansion to use in the display.

col
The color(s) to use for the backgrounds of the boxes containing the category labels.

Details

furc displays one furcation of the dendrogram. A furcation is a single box displaying its label and count that may split into finer divisions. If so, furc calls itself for each furcation until there are no more splits.

On the final level of categorization, the category labels are staggered to minimize overprinting.

Value

nil

Author(s)

Jim Lemon

See Also

plot.dendrite

[Package plotrix version 3.7-7 Index]