get.gantt.info {plotrix}	R Documentation

Gather the information to create a Gantt chart

Description

Allows the user to enter the information for a Gantt chart.

Usage


 get.gantt.info(format="%Y/%m/%d")

Arguments

format
the format to be used in entering dates/times. Defaults to YYYY/mm/dd. See strptime for various date/time formats.

Value

The list used to create the chart. Elements are:

labels
The task labels to be displayed at the left of the chart.

starts,ends
The task starts/ends as POSIXct dates/times.

priorities
Task priorities as integers in the range 1 to 10. There can be less than 10 levels of priority, but if priorities do not start at 1 (assumed to be the highest), the default priority colors will be calculated from 1.

Author(s)

Jim Lemon

See Also

gantt.chart

Examples


 cat("Enter task times using HH:MM (hour:minute) format\n")
 get.gantt.info("%H:%M")

[Package plotrix version 3.8-2 Index]