get.soil.texture {plotrix}	R Documentation

Enter soil texture data

Description

get.soil.texture calls get.triprop to allow the user to enter soil textures as the proportions or percentages of three components, sand, silt and clay.

Usage


 get.soil.texture(use.percentages=FALSE,cnames=c("sand","silt","clay"))

Arguments

use.percentages
Logical - whether to treat the entries as percentages and scale to proportions.

cnames
column names for the resulting three column matrix.

Value

A matrix of the components of one or more soil samples.

Author(s)

Sander Oom and Jim Lemon

See Also

soil.texture,get.triprop

Examples


 if(dev.interactive()) {
  newsp<-get.soil.texture()
  # show the soil triangle
  soil.texture()
  # now plot the observations
  show.soil.texture(newsp)
 }

[Package plotrix version 3.8-2 Index]