getFigCtr {plotrix}	R Documentation

Find the center of the figure region.

Description

Calculates the center of the figure region.

Usage

 getFigCtr()

Details

getFigCtr reads parameters about the current plot and calculates the vertical and horizontal centers of the figure region. This is typically useful for placing a centered title on plots where the left and right margins are very different.

Value

A two element vector containing the coordinates of the center of the figure region in user units.

Author(s)

Jim Lemon

[Package plotrix version 2.7 Index]