getIntersectList {plotrix}	R Documentation

Enter the information for a set intersection display

Description

Enter the information for a set intersection display.

Usage

 getIntersectList(nelem,xnames=NULL)

Arguments

nelem 	 The number of sets for which the intersections will be displayed.

xnames 	 The labels for the set intersections. The function creates names from combinations of the first nelem capital letters if none are given.

Details

getIntersectList allows the user to manually enter the counts of set intersections rather than build this information from a matrix of data. It is probably most useful for producing an intersection diagram when the counts of the intersections are already known.

Value

A list of the counts of elements in the set intersections.

Author(s)

Jim Lemon

See Also

makeIntersectList, intersectDiagram

[Package plotrix version 2.7 Index]