pie.labels {plotrix}	R Documentation

Place labels on a pie chart

Description

Places labels on a pie chart

Usage

 pie.labels(x,y,angles,labels,radius=1,bg="white",border=TRUE,...)

Arguments

x,y 	 x and y position of the center of the pie chart

angles 	 A numeric vector representing angles in radians. This is the return value of floating.pie.

labels 	 Text strings to label each sector.

radius 	 The radius at which to place the labels in user units. The default is 1.

bg 	 The color of the rectangles on which the labels are displayed.

border 	 Whether to draw borders around the rectangles.

... 	 Arguments passed to boxed.labels.

Value

nil

Note

Remember that x and y specify the center of the pie chart and that the label positions are specified by angles and radii from that center.

Author(s)

Jim Lemon

See Also

floating.pie, boxed.labels

Examples

 pieval<-c(2,4,6,8)
 plot(1:5,type="n",axes=FALSE)
 box()
 bisect.angles<-floating.pie(3,3,pieval)
 pie.labels(3,3,bisect.angles,c("two","four","six","eight"))

[Package plotrix version 2.7 Index]