plot.dendrite {plotrix}	R Documentation

Plot a dendrogram of a dendrite object

Description

Plot a dendrogram for two or more mutually exclusive attributes.

Usage

 ## S3 method for class 'dendrite'
plot(x,xlabels=NULL,main="",mar=c(1,0,3,0),cex=1,
  col="white",...)

Arguments

x
A dendrite object containing the counts of objects having combinations of mutually exclusive attributes.

xlabels
The category labels that will be displayed beneath the dendrogram.

main
The title of the plot.

mar
Margins for the plot.

cex
Character expansion for the leaves of the dendrogram.

col
Background colors for the boxes (see Details).

...
Additional arguments passed to plot.

Details

plot.dendrite sets up a plot for a dendrogram. The actual plotting of the dendrogram is done by furc.

If sampcol is a vector of colors as long as the number of category labels, each category label will be displayed in a different colored box. This may may cause problems when different category labels are used at each level with common names, as the color for a name will be the same at all levels.

Value

nil

Author(s)

Jim Lemon

See Also

furc

Examples

 sex<-sample(c("M","F"),100,TRUE)
 hair<-c(sample(c("Blond","Black","Brown","Red"),90,TRUE),rep(NA,10))
 eye<-sample(c("Blue","Black","Brown","Green"),100,TRUE)
 charac<-data.frame(sex=sex,hair=hair,eye=eye)
 characlist<-makeDendrite(charac)
 shecol<-c(Black="black",Blond="yellow",Blue="blue",Brown="brown",
  F="pink",Green="green",M="lightblue","NA"="gray",Red="orange")
 plot.dendrite(characlist,names(charac),main="Test dendrogram",
  col=shecol)

[Package plotrix version 3.7-7 Index]