plot_bg {plotrix}	R Documentation

Add a background color to a plot

Description

Displays a colored rectangle over the entire area of a plot

Usage

plot_bg(col="lightgray")

Arguments

col
The color of the background

Details

plot_bg is probably only useful when part of the do.first argument in another plot function to add a background color to the plot.

Value

nil

Author(s)

Jim Lemon

Examples


 barp(1:5,do.first="plot_bg()",col=1:5)

[Package plotrix version 3.8-2 Index]