print.brklist {plotrix}	R Documentation

Display the output of brkdnNest

Description

Displays the list of values produced by brkdnNest.

Usage


 ## S3 method for class 'brklist'
print(x,...)

Arguments

x
a list of summary values produced by \link{brkdnNest}

...
additional arguments passed to print.

Details

print.brklist displays frequency tables produced by brkdnNest. It is mainly for convenience, but does make a nicer display than when passed directly to print

Value

nil

Author(s)

Jim Lemon

See Also

brkdnNest

Examples


 printbrktest<-data.frame(A=c(sample(1:10,99,TRUE),NA),
  B=sample(c("Yes","No"),100,TRUE),
  C=sample(LETTERS[1:3],100,TRUE))
 pbt<-brkdnNest(A~B+C,printbrktest)
 print(pbt)

[Package plotrix version 3.8-2 Index]