sliceArray {plotrix}	R Documentation

Slice an array

Description

Slices one dimension from an array by taking one element from the first dimension.

Usage


 sliceArray(x,slice)

Arguments

x
An array

slice
The index of the slice to take from the first dimension of the array.

Details

sliceArray builds an extractor string containing the value of slice as the first element and as many commas as needed to match the dimensions of the array. It then applies the extractor function to x and returns the result. Note how the array "slice" swaps dimensions in the example.

Value

The desired slice of the array.

Author(s)

Jim Lemon

See Also

array

Examples


 a1<-array(1:27,dim=c(3,3,3))
 a1
 sliceArray(a1,2)

[Package plotrix version 3.8-2 Index]