sumDendrite {plotrix}	R Documentation

Sum the counts in the top level of a dendrite object.

Description

Find the sum of the counts that are the first elements of each list in the top level of a dendrite object.

Usage

 sumDendrite(x) 

Arguments

x 	 A list with a numeric value as the first element in each of its toplevel elements.

Details

A dendrite object is a possibly nested list of lists that contain the counts and pointers to sublists in each list. Such an object describes the attributes of objects that can take on mutually exclusive attributes (that is, belong to disjunct sets). sumDendrite is a convenience function to get the total number of objects that are so classified.

Value

The sum of the counts in the top level of lists.

Author(s)

Jim Lemon

See Also

plot.dendrite

[Package plotrix version 2.7 Index]