sumbrk {plotrix}	R Documentation

Count specified values in a vector

Description

Counts the number of values in a vector that are equal to a specified value.

Usage


 sumbrk(x,trueval=TRUE,na.rm=TRUE)

Arguments

x
a character, factor or numeric vector.

trueval
the value to be matched in x.

na.rm
whether to remove NA values.

Details

sumbrk counts the values in x matching a specified value. It is mainly to allow these sums to be calculated in the brkdnNest function.

Value

nil

Author(s)

Jim Lemon

See Also

brkdnNest

Examples


 sumbrk(sample(LETTERS,100,TRUE),trueval="M")

[Package plotrix version 3.8-2 Index]