symbolbox {plotrix}	R Documentation

Draw a box filled with symbols

Description

Draws a box on the current figure that is filled with symbols representing individual counts

Usage


 symbolbox(x1,y1,x2,y2,tot,relw=0.5,fg=par("fg"),bg=par("bg"),box=TRUE,
  debug = TRUE,...)

Arguments

x1
left side of box

y1
bottom side of box

x2
right side of box

y2
top side of box

tot
total number of symbols to put in the box

relw
relative width (relative to height) of symbols

fg
foreground color

bg
background color

box
(logical) draw box border?

debug
debug output?

...
additional arguments to polygon() for drawing box

Details

tries to automatically figure out appropriate scaling to fit symbols into the box

Value

none; draws on the current figure

Author(s)

Ben Bolker

See Also

multsymbolbox

Examples


 plot(1:10,1:10,type="n")
 symbolbox(2,5,3,7,tot=20)
 symbolbox(6,2,10,6,tot=50,fg="blue",bg="magenta")

[Package plotrix version 3.8-2 Index]